% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsdataanalysis.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state
Map State Motor Vehicle Fatalities.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{is a numerical code for US state.}

\item{year}{as an integer}
}
\value{
plot of Selected US STATE with all the fatalities in the graphic window
}
\description{
\code{fars_map_state} maps state motor vehicle fatalities given a year and
  state id number.
}
\details{
This function takes a state number and a year, and draws
  a state outline with dots to represent the location of motor vehicle
  fatalities for that year.  An error message will be displayed if an invalid
  state number is chosen or the chosen year's data does not exist.

It require to load maps library (which will load also all the dependacies:mapproj, mapdata, sp, maptools,rnaturalearth)
 It require to load also the graphics library with it's dependacies
}
\examples{
\dontrun{
fars_map_state(12, 2014)
}

}
