% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsdataanalysis.R
\name{fars_read}
\alias{fars_read}
\title{fars_read
This function reads data from the  US National Highway Traffic Safety
Administration's Fatality Analysis Reporting System (FARS).
It proceed by reading the filename of the data povided as argument and return a data as a tibble.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string pointing to the filename of the FARS data.}
}
\value{
This function returns a tibble containing the FARS data. The function will stop if invalide filename is provided.
}
\description{
fars_read
This function reads data from the  US National Highway Traffic Safety
Administration's Fatality Analysis Reporting System (FARS).
It proceed by reading the filename of the data povided as argument and return a data as a tibble.
}
\note{
this package comes with three example files that can be accessed using
fullname <- system.file('extdata', 'accident_2013.csv.bz2', package = 'farsdataanalysis')
fars_read(filename = fullname)
}
\examples{
\dontrun{
fars_read("accident_2015.csv.bz2")
}

}
