% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsdataanalysis.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Vector of years' FARS files to open.
Valid vector element will be integer,string or numeric that can be coerced to integer.}
}
\value{
This function returns a summureze tibble in a wide form
}
\description{
Produce a Summary of FARS Files: number of fatalities per mounth per year.
}
\details{
\code{fars_summarize_years} produces a summary tibble of FARS years and
  months given a vector of years.

This function takes a vector of years, extract the FARS data for the corresponding years.
  Bind the tibbles in the data list together. the function dplyr::bind_rows() allow to combine the elements (tibbles) in dat_list
  with differing number of variables together to create just one combined dataset.
  Produce the summary of combined dataset(tribble) showing the number of observations as per combination MONTH/year.
  This function require a valid input argument: i.e. a valid years vector as requiered for FARS data.
}
\examples{
\dontrun{
fars_summarize_years(years = c(2013,2014,2015))
}
}
