% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsdataanalysis.R
\name{make_filename}
\alias{make_filename}
\title{make_filename.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer,a string or a numeric data type that can be coerced to an integer
But this function can be improve by adding argument validation test, that will check if the input argument
can be coerce to integer. If yes continue with coercion. Else ,request the users to provide a valid input argument.}
}
\value{
this function returns a string that is the proper FARS data
   filename for the given year.  An erronous file name can be produce the param year was not validate.
}
\description{
\code{make_filename} makes a file name by adding the given year as per  FARS filename standard
}
\details{
This function takes a year as input argument and produces a valid FARS filename as output.
in this format: "accident_" + year + ".csb.bz2"
}
\examples{
\dontrun{
make_filename(2013)
}

}
